backend/requirements.txt
# FastAPI Backend for PortfolioPath
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Data & Finance
yfinance>=0.2.36
pandas>=2.1.4
numpy>=1.26.3

# Caching
redis>=5.0.1
aioredis>=2.0.1

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# CORS
starlette>=0.35.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
